+++ YNDropDownMenu/Classes/YNDropDownDelegate.swift
@@ -11,10 +11,4 @@ import Foundation
 public protocol YNDropDownDelegate: class {
     func hideMenu()
 }
-open class YNDropDownView: UIView {
-    open weak var delegate: YNDropDownDelegate?
-    
-    open func hideMenu() {
-        self.delegate?.hideMenu()
-    }
-}
+