+++ CHANGELOG.md
@@ -5,9 +5,13 @@ it, simply add the following line to your Podfile:
 
 `pod 'YNDropDownMenu'`
 
+## [v1.3.0](https://github.com/younatics/YNDropDownMenu/releases/tag/1.3.0)
+* Override method added to find drop down view open&close
+* Init method `YNDropDownMenu(frame: frame, YNDropDownViews: [YNDropDownView], dropDownViewTitles: [String])` deprecated
+* Add `view.alwaysSelectedAt(index: index)` function
+
 ## [v1.2.3](https://github.com/younatics/YNDropDownMenu/releases/tag/1.2.3)
 * Add Image Folder
-* Edtited readme.md
 
 ## [v1.2.2](https://github.com/younatics/YNDropDownMenu/releases/tag/1.2.2)
 * Init style added UIView and YNDropDownView